\name{BRCA}
\alias{BRCA}
\docType{data}
\title{
Postive and negative estrogen receptor status with breast cancer data
}
\description{
Includes the gene expressiong data samples only for breast cancer data
}
\usage{data("BRCA")}
\examples{
data(BRCA)
## maybe str(BRCA) ; plot(BRCA) ...
}
\keyword{datasets}
