\name{subBRCA}
\alias{subBRCA}
\docType{data}
\title{
Postive and negative estrogen receptor status with breast cancer data
}
\description{
Includes the samples that has postive and negative estrogen receptor status with breast cancer data
}
\usage{data("subBRCA")}

\examples{
data(subBRCA)

}
\keyword{datasets}
